requirements.txt
streamlit
google-genai
python-dotenv
pydantic
pypdf
python-docx
reportlab